changing decimation example to all embedded filters example
