+++ examples/embedded-filters/README.md
@@ -1,0 +1,13 @@
+# Embedded Filters Example
+
+This example demonstrates how to use the `rs2::embedded_filter` class for RealSense devices that support embedded filters.
+
+## Scenario
+
+The example shows how to:
+1. Find embedded filters supported by the first found connected RealSense device
+2. Read current options for each supported filter
+3. Modify various options for each supported filter
+4. Setting back the options' to their initial value
+
+
